refactor: rename to updater
