+++ lib/views/home/home-screen.dart
@@ -3,8 +3,8 @@ import 'dart:async';
 import 'package:fab_circular_menu/fab_circular_menu.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
+import 'package:syphon/global/libs/updater/update-check.dart';
 import 'package:syphon/global/strings.dart';
-import 'package:syphon/global/update-check.dart';
 import 'package:syphon/store/hooks.dart';
 import 'package:syphon/store/index.dart';
 import 'package:syphon/store/rooms/room/model.dart';
@@ -31,9 +31,8 @@ class HomeScreen extends HookWidget {
       (state) => state.syncStore.lastSince,
     );
 
-    final checkForUpdatesEnabled = useSelector<AppState, bool>(
-            (state) => state.settingsStore.checkForUpdatesEnabled) ??
-        false;
+    final checkForUpdatesEnabled =
+        useSelector<AppState, bool>((state) => state.settingsStore.checkForUpdatesEnabled) ?? false;
 
     final searchLoading = useSelector<AppState, bool>(
           (state) => state.searchStore.loading,
@@ -45,13 +44,13 @@ class HomeScreen extends HookWidget {
         ) ??
         [];
 
-    final fabType = useSelector<AppState, MainFabType>(
-            (state) => state.settingsStore.themeSettings.mainFabType) ??
-        MainFabType.Ring;
+    final fabType =
+        useSelector<AppState, MainFabType>((state) => state.settingsStore.themeSettings.mainFabType) ??
+            MainFabType.Ring;
 
-    final fabLabels = useSelector<AppState, MainFabLabel>(
-            (state) => state.settingsStore.themeSettings.mainFabLabel) ??
-        MainFabLabel.Off;
+    final fabLabels =
+        useSelector<AppState, MainFabLabel>((state) => state.settingsStore.themeSettings.mainFabLabel) ??
+            MainFabLabel.Off;
 
     final fabLocation = useSelector<AppState, MainFabLocation>(
             (state) => state.settingsStore.themeSettings.mainFabLocation) ??
